+++ com-psi/OblvPermutation.cpp
@@ -12,7 +12,7 @@ namespace osuCrypto
         Matrix<u8> src,
         std::string tag)
     {
-        u32 rows = src.rows();
+        u32 rows = gsl::narrow<u32>(src.rows());
         //std::vector<u32> pi1(src.rows());
         //for (u32 i = 0; i < pi1.size(); ++i)
         //{
@@ -102,7 +102,7 @@ namespace osuCrypto
         std::string tag,
         OutputType type)
     {
-        u32 recvCount = (srcRows + step - 1) / step;
+        u32 recvCount = gsl::narrow<u32>((srcRows + step - 1) / step);
 
         std::vector<std::pair<std::future<void>, std::vector<u8>>> recvs1(recvCount);
         std::vector<std::pair<std::future<void>, std::vector<u32>>> recvs2(recvCount);
@@ -182,7 +182,7 @@ namespace osuCrypto
         auto permPointer = std::make_shared<std::vector<u32>>(std::move(permutation));
 
         auto& perm = *permPointer;
-        auto rows = perm.size();
+        auto rows = gsl::narrow<u32>(perm.size());
 
         //#ifndef NDEBUG
         //        if (true)
@@ -228,7 +228,7 @@ namespace osuCrypto
                 auto buffer = prng.getBufferSpan(min * sizeof(u32));
                 auto u32Buffer = span<u32>((u32*)buffer.data(), buffer.size() / sizeof(u32));
 
-                min -= u32Buffer.size();
+                min -= (u32)(u32Buffer.size());
 
                 for (u64 k = 0; k < u32Buffer.size(); ++k)
                 {
